% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KenRCATs_helpers.R
\name{append_KenRCATs}
\alias{append_KenRCATs}
\title{Append KenRCATs Results with Input Parameters}
\usage{
append_KenRCATs(df, input)
}
\arguments{
\item{df}{A data frame containing simulation results.}

\item{input}{A list of input parameters from the Shiny app.}
}
\value{
A list containing two data frames: the original results and the results
with appended input parameters.
}
\description{
Append KenRCATs Results with Input Parameters
}
\keyword{internal}
