% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_diagnostics.R
\name{cluster_diag_cluster_fits}
\alias{cluster_diag_cluster_fits}
\title{Cluster-Specific Linear Fits}
\usage{
cluster_diag_cluster_fits(dat)
}
\description{
Cluster-Specific Linear Fits
}
\keyword{internal}
