% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_diagnostics.R
\name{cluster_diag_extract_weights}
\alias{cluster_diag_extract_weights}
\title{Extract Robust Regression Weights}
\usage{
cluster_diag_extract_weights(fit, n)
}
\description{
Extract Robust Regression Weights
}
\keyword{internal}
