% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_diagnostics.R
\name{cluster_diag_method_labels}
\alias{cluster_diag_method_labels}
\title{Diagnostic Method Labels}
\usage{
cluster_diag_method_labels(methods)
}
\description{
Diagnostic Method Labels
}
\keyword{internal}
