% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_diagnostics.R
\name{cluster_diag_observation_diagnostics}
\alias{cluster_diag_observation_diagnostics}
\title{Observation-Level OLS and Robust Diagnostics}
\usage{
cluster_diag_observation_diagnostics(dat)
}
\description{
Observation-Level OLS and Robust Diagnostics
}
\keyword{internal}
