% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_diagnostics.R
\name{cluster_diag_prepare_data}
\alias{cluster_diag_prepare_data}
\title{Prepare Canonical Diagnostic Data}
\usage{
cluster_diag_prepare_data(data, response, predictor, cluster)
}
\description{
Prepare Canonical Diagnostic Data
}
\keyword{internal}
