% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_diagnostics.R
\name{cluster_diag_validate_specification}
\alias{cluster_diag_validate_specification}
\title{Validate a Cluster Diagnostic Specification}
\usage{
cluster_diag_validate_specification(formula, cluster, data)
}
\description{
Validate a Cluster Diagnostic Specification
}
\keyword{internal}
