% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KenRCATs_helpers.R
\name{getUIParams}
\alias{getUIParams}
\title{Generate UI Parameters for KenRCATs Shiny App}
\usage{
getUIParams()
}
\value{
A shiny::tagList containing UI elements.
}
\description{
This internal function creates a list of UI elements for the KenRCATs() 'shiny'
application. It generates input fields for various parameters used in the power
analysis simulation.
}
\note{
This function is intended for internal use within the KenRCATs() 'shiny'
application.
}
\keyword{internal}
