% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kenward_roger.R
\name{kenward_roger}
\alias{kenward_roger}
\title{Kenward-Roger Analysis Shiny Application}
\usage{
kenward_roger()
}
\value{
A Shiny application object.
}
\description{
A Shiny application that allows users to upload a dataset, modify variable types,
and fit a mixed-effects model using the Kenward-Roger approximation for small
sample inference.
}
\details{
The application provides an interactive interface for setting up and running
a mixed-effects model analysis with the Kenward-Roger method for estimating
the degrees of freedom in linear mixed-effects models.
}
\examples{
if (interactive()) {
  kenward_roger()
}

}
\references{
Kenward, M. G., & Roger, J. H. (1997). Small Sample Inference for Fixed Effects
from Restricted Maximum Likelihood. Biometrics, 53, 983-997.

Kuznetsova, A., Brockhoff, P. B., & Christensen, R. H. B. (2017). lmerTest
Package: Tests in Linear Mixed Effects Models. Journal of Statistical Software,
82(13), 1-26. <doi: 10.18637/jss.v082.i13>.
}
