% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KenRCATs_helpers.R
\name{list_null}
\alias{list_null}
\title{Handle Null Values for List Input}
\usage{
list_null(par_input = "")
}
\arguments{
\item{par_input}{A string input, default is "".}
}
\value{
NULL if input is NA or empty, otherwise a parsed list.
}
\description{
Handle Null Values for List Input
}
\keyword{internal}
