% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KenRCATs_helpers.R
\name{parse_list_input}
\alias{parse_list_input}
\title{Parse String Input to List}
\usage{
parse_list_input(input_string)
}
\arguments{
\item{input_string}{A string representation of a list.}
}
\value{
A parsed list object.
}
\description{
Parse String Input to List
}
\keyword{internal}
