% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_game_mgmt.R
\name{render_card_grid}
\alias{render_card_grid}
\title{Render Card Grid in Shiny App}
\usage{
render_card_grid(new_card_grid)
}
\arguments{
\item{new_card_grid}{A matrix or data frame where each row represents a card and
each card has a property \code{icard} pointing to the image file relative to the
\code{mmiCATs} package's \code{www} directory. The function expects this parameter to be
structured with named columns where \code{icard} is one of the column names.}
}
\value{
A Shiny UI element (\code{tagList}) representing a grid of card images.
}
\description{
This function takes a grid of card information, generates image tags for
each card, and organizes them into a responsive grid layout for display in a
Shiny application.
}
\keyword{internal}
