% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KenRCATs_helpers.R
\name{sig_null}
\alias{sig_null}
\title{Handle Null Values for Sigma Matrix}
\usage{
sig_null(par_input = "")
}
\arguments{
\item{par_input}{A string input, default is "".}
}
\value{
NULL if input is NA or empty, otherwise an evaluated expression.
}
\description{
Handle Null Values for Sigma Matrix
}
\keyword{internal}
