% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwr_func_study1_helpers.R
\name{study1_binomial_mcse}
\alias{study1_binomial_mcse}
\title{Calculate a Binomial Monte Carlo Standard Error}
\usage{
study1_binomial_mcse(proportion, n)
}
\arguments{
\item{proportion}{Estimated probability.}

\item{n}{Number of successful replications.}
}
\value{
A scalar numeric value.
}
\description{
Calculate a Binomial Monte Carlo Standard Error
}
\keyword{internal}
