% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwr_func_study1_helpers.R
\name{study1_check_integer}
\alias{study1_check_integer}
\title{Check an Integer Scalar}
\usage{
study1_check_integer(x, name, lower = -Inf, upper = Inf)
}
\arguments{
\item{x}{Object to check.}

\item{name}{Argument name.}

\item{lower}{Optional lower bound.}

\item{upper}{Optional upper bound.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Check an Integer Scalar
}
\keyword{internal}
