% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwr_func_study1_helpers.R
\name{study1_collapse_messages}
\alias{study1_collapse_messages}
\title{Collapse Warning or Error Messages}
\usage{
study1_collapse_messages(messages)
}
\arguments{
\item{messages}{Character vector of messages.}
}
\value{
A single character value or \code{NA_character_}.
}
\description{
Collapse Warning or Error Messages
}
\keyword{internal}
