% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwr_func_study1_helpers.R
\name{study1_extract_cats}
\alias{study1_extract_cats}
\title{Extract a CATs Result}
\usage{
study1_extract_cats(cats_fit, alpha, n_clusters)
}
\arguments{
\item{cats_fit}{CATs output.}

\item{alpha}{Significance level.}

\item{n_clusters}{Maximum number of clusters.}
}
\value{
A standardized result list.
}
\description{
Extract a CATs Result
}
\keyword{internal}
