% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwr_func_study1_helpers.R
\name{study1_fit_cats}
\alias{study1_fit_cats}
\title{Fit an Ordinary CATs Method}
\usage{
study1_fit_cats(dat, alpha, truncate)
}
\arguments{
\item{dat}{Simulated data.}

\item{alpha}{Significance level.}

\item{truncate}{Whether to drop outlying cluster-specific estimates.}
}
\value{
A standardized result list.
}
\description{
Fit an Ordinary CATs Method
}
\keyword{internal}
