% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwr_func_study1_helpers.R
\name{study1_fit_cr2}
\alias{study1_fit_cr2}
\title{Fit the CR2 Benchmark}
\usage{
study1_fit_cr2(dat, alpha)
}
\arguments{
\item{dat}{Simulated data.}

\item{alpha}{Significance level.}
}
\value{
A standardized result list.
}
\description{
Fit the CR2 Benchmark
}
\keyword{internal}
