% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwr_func_study1_helpers.R
\name{study1_fit_robust_cats}
\alias{study1_fit_robust_cats}
\title{Fit a Robust CATs Method}
\usage{
study1_fit_robust_cats(dat, alpha, engine)
}
\arguments{
\item{dat}{Simulated data.}

\item{alpha}{Significance level.}

\item{engine}{Robust regression engine.}
}
\value{
A standardized result list.
}
\description{
Fit a Robust CATs Method
}
\keyword{internal}
