% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwr_func_study1_helpers.R
\name{study1_mean_or_na}
\alias{study1_mean_or_na}
\title{Calculate a Mean or Return Missing}
\usage{
study1_mean_or_na(x)
}
\arguments{
\item{x}{Numeric or logical vector.}
}
\value{
A scalar numeric value.
}
\description{
Calculate a Mean or Return Missing
}
\keyword{internal}
