% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwr_func_study1_helpers.R
\name{study1_method_names}
\alias{study1_method_names}
\title{Study 1 Method Names}
\usage{
study1_method_names()
}
\value{
A character vector of method names.
}
\description{
Returns the method names supported by \code{pwr_func_study1()}.
}
\keyword{internal}
