% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwr_func_study1_helpers.R
\name{study1_rmse_or_na}
\alias{study1_rmse_or_na}
\title{Calculate RMSE or Return Missing}
\usage{
study1_rmse_or_na(estimate, truth)
}
\arguments{
\item{estimate}{Estimated coefficients.}

\item{truth}{True coefficients.}
}
\value{
A scalar numeric value.
}
\description{
Calculate RMSE or Return Missing
}
\keyword{internal}
