% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwr_func_study1_helpers.R
\name{study1_summarize_results}
\alias{study1_summarize_results}
\title{Summarize Study 1 Replicate Results}
\usage{
study1_summarize_results(replicate_results, methods, reps)
}
\arguments{
\item{replicate_results}{Replicate-level results.}

\item{methods}{Ordered method names.}

\item{reps}{Number of requested replications.}
}
\value{
A data frame of method-level simulation summaries.
}
\description{
Summarize Study 1 Replicate Results
}
\keyword{internal}
