% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwr_func_study2_helpers.R
\name{study2_method_names}
\alias{study2_method_names}
\title{Study 2 Method Names}
\usage{
study2_method_names()
}
\value{
A character vector of method names.
}
\description{
Returns the method names supported by \code{pwr_func_study2()}.
}
\keyword{internal}
