% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_mixed_models.R
\name{study_require_robustlmm}
\alias{study_require_robustlmm}
\title{Require the Robust Mixed-Model Dependency}
\usage{
study_require_robustlmm()
}
\value{
\code{NULL}, invisibly.
}
\description{
Require the Robust Mixed-Model Dependency
}
\keyword{internal}
