% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KenRCATs_helpers.R
\name{text_to_vector}
\alias{text_to_vector}
\title{Convert Text Input to Vector}
\usage{
text_to_vector(text_input)
}
\arguments{
\item{text_input}{A string input to be converted to a vector.}
}
\value{
A vector parsed from the input string.
}
\description{
Convert Text Input to Vector
}
\keyword{internal}
