% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KenRCATs_helpers.R
\name{vec_null}
\alias{vec_null}
\title{Handle Null Values for Text to Vector Conversion}
\usage{
vec_null(par_input = "")
}
\arguments{
\item{par_input}{A string input, default is "".}
}
\value{
NULL if input is NA or empty, otherwise a vector.
}
\description{
Handle Null Values for Text to Vector Conversion
}
\keyword{internal}
