% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_im_lmRob.R
\name{fit_model}
\alias{fit_model}
\title{Fit a Linear Model with Robust Estimation}
\usage{
fit_model(engine, formula, data, ...)
}
\arguments{
\item{engine}{A character string specifying the engine to be used for model
fitting. Must be either "robust" or "robustbase".}

\item{formula}{An object of class formula (or one that can be coerced to that
class): a symbolic description of the model to be fitted.}

\item{data}{A dataframe}

\item{...}{Additional arguments to be passed to the underlying fitting function.}
}
\value{
A fitted model.
}
\description{
This function fits a linear model using robust estimation methods.
It allows the use of either the 'robust' or 'robustbase' packages for fitting
the model.
}
\keyword{internal}
